ADD: Merging process...
