+++ src/lottery/index.js
@@ -22,6 +22,7 @@ lotteryRouter.use("/global-state", require("./routes/globalState"));
 lotteryRouter.use("/transactions", require("./routes/transactions"));
 lotteryRouter.use("/items", require("./routes/items"));
 lotteryRouter.use("/public-notice", require("./routes/publicNotice"));
+lotteryRouter.use("/quests", require("./routes/quests"));
 
 const resources = [
   eventStatusModel,
