+++ src/lottery/modules/populates/publicNotice.js
@@ -1,0 +1,11 @@
+const publicNoticePopulateOption = [
+  { path: "event" },
+  {
+    path: "item",
+    select: "name price description itemType",
+  },
+];
+
+module.exports = {
+  publicNoticePopulateOption,
+};
