+++ src/lottery/routes/publicNotice.js
@@ -3,8 +3,8 @@ const express = require("express");
 const router = express.Router();
 const publicNoticeHandlers = require("../services/publicNotice");
 
-// 상점이용은 로그인을 요구합니다.
-router.use(require("../../middlewares/auth"));
-router.get("/get-recent-transaction", publicNotice.getRecentTransaction);
+// 상점 공지는 로그인을 요구하지 않습니다.
+router.get("/recent-transactions", publicNoticeHandlers.getRecentTransaction);
+router.get("/leaderboard", publicNoticeHandlers.getTicketLeaderboardHandler);
 
 module.exports = router;
